Refactor: Bitline module structure
Split the large bitline.rs to make it easier to read.
Move the trait definition of Bitline to base.rs.
Move the part related to the specific implementation of u8, u16, ... to uints.rs.
Add tests to ensure that the external interface has not changed. no breaking changes
