+++ src/bitline/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod uints;
+// re-export
+pub use base::Bitline;
+pub use uints::{Bitline128, Bitline16, Bitline32, Bitline64, Bitline8};
